Add a unique id to each request
Under high load, it is possible that thread A enqueues a request before thread B to
the same shard worker within a very short interval of time. Both threads go to sleep
blocking on results. The result arrive in the response queue in order but thread B
wakes up before thread A and receives the response meant for A.

To naively fix the issue enqueue a unique id for each request, and the worker returns
the unique id along with the response. The request process retries multiple times to
find the response with the correct id in the queue.